[Feat]: my info ui 구현
