+++ .idea/assetWizardSettings.xml
@@ -34,8 +34,8 @@
                         <option name="values">
                           <map>
                             <entry key="assetSourceType" value="FILE" />
-                            <entry key="outputName" value="chevron_down_small_line" />
-                            <entry key="sourceFile" value="$USER_HOME$/icon_list/nav/chevron_down_small_line.svg" />
+                            <entry key="outputName" value="close_24px" />
+                            <entry key="sourceFile" value="$USER_HOME$/icon_list/12-31/navigation/close_24px.svg" />
                           </map>
                         </option>
                       </PersistentState>
