+++ feature/myaccount/src/main/java/com/android/myfooddiarybookaos/myaccount/myInfo/MyInfoScreen.kt
@@ -1,0 +1,233 @@
+package com.android.myfooddiarybookaos.myaccount.myInfo
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Divider
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavHostController
+import com.android.myfooddiarybookaos.core.data.R
+import com.android.myfooddiarybookaos.data.TextBox
+import com.android.myfooddiarybookaos.data.component.coloredInnerShadow
+import com.android.myfooddiarybookaos.data.robotoBold
+import com.android.myfooddiarybookaos.data.robotoRegular
+import com.android.myfooddiarybookaos.data.utils.scaledSp
+
+@Composable
+fun MyInfoScreen(
+    myNavi: NavHostController,
+) {
+
+    val isBackState = remember { mutableStateOf(false) }
+    val passwordChangeState = remember { mutableStateOf(false) }
+    val allDiaryDeleteState = remember { mutableStateOf(false) }
+    val logoutState = remember { mutableStateOf(false) }
+    val deleteUserState = remember { mutableStateOf(false) }
+
+    if (isBackState.value) {
+        myNavi.popBackStack()
+        isBackState.value = false
+    }
+
+    Column {
+        Column(
+            modifier = Modifier
+                .height(90.dp)
+                .fillMaxWidth(),
+            verticalArrangement = Arrangement.Bottom
+        ) {
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(start = 4.dp)
+            ) {
+                Box(
+                    modifier = Modifier
+                        .size(44.dp)
+                        .align(Alignment.BottomStart)
+                        .padding(bottom = 2.dp)
+                        .clickable(onClick = { isBackState.value = true }),
+                    contentAlignment = Alignment.Center,
+                ) {
+                    Image(
+                        painter = painterResource(id = R.drawable.chevron_left_24px),
+                        contentDescription = "",
+                    )
+                }
+
+                Box(
+                    modifier = Modifier
+                        .wrapContentSize()
+                        .align(Alignment.BottomCenter)
+                        .padding(14.75.dp)
+                ) {
+                    TextBox(
+                        text = "내 정보",
+                        fontWeight = 500,
+                        fontFamily = robotoBold,
+                        fontSize = 18.scaledSp(),
+                        color = colorResource(id = R.color.black)
+                    )
+                }
+
+            }
+
+            Spacer(modifier = Modifier.height(1.dp))
+
+            Divider(
+                modifier = Modifier
+                    .height(2.dp)
+                    .coloredInnerShadow(
+                        color = colorResource(id = R.color.black_10),
+                        offsetY = 1.dp,
+                        blurRadius = 4.dp
+                    )
+            )
+
+
+        }
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .fillMaxWidth()
+                .clickable {
+                    passwordChangeState.value = true
+                }
+                .padding(10.dp)
+        ) {
+            Box(
+                modifier = Modifier.size(24.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.create_24px),
+                    contentDescription = null
+                )
+            }
+
+            Spacer(modifier = Modifier.width(10.dp))
+
+            Text(
+                text = "비밀번호 변경",
+                fontFamily = robotoRegular,
+                fontSize = 20.sp,
+                fontWeight = FontWeight.W400,
+                lineHeight = 24.sp,
+                color = Color.Black
+            )
+        }
+
+        Spacer(modifier = Modifier.height(34.dp))
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .fillMaxWidth()
+                .clickable {
+                    allDiaryDeleteState.value = true
+                }
+                .padding(10.dp)
+        ) {
+            Box(
+                modifier = Modifier.size(24.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.component_1),
+                    contentDescription = null
+                )
+            }
+
+            Spacer(modifier = Modifier.width(10.dp))
+
+            Text(
+                text = "모든 식사일기 삭제",
+                fontFamily = robotoRegular,
+                fontSize = 20.sp,
+                fontWeight = FontWeight.W400,
+                lineHeight = 24.sp,
+                color = Color.Black
+            )
+        }
+
+        Spacer(modifier = Modifier.height(34.dp))
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .fillMaxWidth()
+                .clickable {
+                    deleteUserState.value = true
+                }
+                .padding(10.dp)
+        ) {
+            Box(
+                modifier = Modifier.size(24.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.close_24px),
+                    contentDescription = null
+                )
+            }
+
+            Spacer(modifier = Modifier.width(10.dp))
+
+            Text(
+                text = "회원탈퇴",
+                fontFamily = robotoRegular,
+                fontSize = 20.sp,
+                fontWeight = FontWeight.W400,
+                lineHeight = 24.sp,
+                color = Color.Black
+            )
+        }
+
+        Spacer(modifier = Modifier.height(34.dp))
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .fillMaxWidth()
+                .clickable {
+                    logoutState.value = true
+                }
+                .padding(10.dp)
+        ) {
+            Box(
+                modifier = Modifier.size(24.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.component_2),
+                    contentDescription = null
+                )
+            }
+
+            Spacer(modifier = Modifier.width(10.dp))
+
+            Text(
+                text = "로그아웃",
+                fontFamily = robotoRegular,
+                fontSize = 20.sp,
+                fontWeight = FontWeight.W400,
+                lineHeight = 24.sp,
+                color = Color.Black
+            )
+        }
+
+    }
+
+}
