+++ feature/myaccount/src/main/java/com/android/myfooddiarybookaos/myaccount/navi/MyGraph.kt
@@ -4,9 +4,9 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.android.myfooddiarybookaos.data.state.ApplicationState
+import com.android.myfooddiarybookaos.myaccount.myInfo.MyInfoScreen
 import com.android.myfooddiarybookaos.myaccount.myMain.MyMainScreen
-import com.android.myfooddiarybookaos.myaccount.notice.NoticeScreen
+import com.android.myfooddiarybookaos.myaccount.myNotice.NoticeScreen
 
 @Composable
 fun MyGraph(
@@ -24,6 +24,10 @@ fun MyGraph(
 
         composable(MyScreenRoot.NOTICE){
             NoticeScreen(myNavi)
+        }
+
+        composable(MyScreenRoot.INFO){
+            MyInfoScreen(myNavi)
         }
     }
 }
