+++ feature/myaccount/src/main/java/com/android/myfooddiarybookaos/myaccount/navi/MyScreenRoot.kt
@@ -3,4 +3,5 @@ package com.android.myfooddiarybookaos.myaccount.navi
 object MyScreenRoot {
     const val MY = "myMainScreen"
     const val NOTICE = "noticeScreen"
+    const val INFO = "infoScreen"
 }
